+++ Source/DataSources/ModelVisualizer.js
@@ -20,6 +20,7 @@ define([
         ModelAnimationLoop,
         Property) {
     "use strict";
+    /*global console*/
 
     var defaultScale = 1.0;
     var defaultMinimumPixelSize = 0.0;
@@ -105,7 +106,7 @@ define([
                     url : uri
                 });
 
-                model.readyToRender.addEventListener(readyToRender, this);
+                model.readyPromise.then(onModelReady).otherwise(onModelError);
 
                 model.id = entity;
                 primitives.add(model);
@@ -185,17 +186,20 @@ define([
     function removeModel(visualizer, entity, modelHash, primitives) {
         var modelData = modelHash[entity.id];
         if (defined(modelData)) {
-            var model = modelData.modelPrimitive;
-            model.readyToRender.removeEventListener(readyToRender, visualizer);
-            primitives.removeAndDestroy(model);
+            primitives.removeAndDestroy(modelData.modelPrimitive);
             delete modelHash[entity.id];
         }
     }
 
-    function readyToRender(model) {
+    function onModelReady(model) {
         model.activeAnimations.addAll({
             loop : ModelAnimationLoop.REPEAT
         });
     }
+
+    function onModelError(error) {
+        console.error(error);
+    }
+
     return ModelVisualizer;
 });